+++ packages/money-account-upgrade-controller/src/steps/step.ts
@@ -10,6 +10,7 @@ export type StepContext = {
   address: Hex;
   chainId: Hex;
   delegateAddress: Hex;
+  delegationManager: Hex;
   delegatorImplAddress: Hex;
   erc20TransferAmountEnforcer: Hex;
   musdTokenAddress: Hex;